+++ apps/lite/src/app/core/guards/auth/auth.guard.ts
@@ -1,18 +1,11 @@
 import { Injectable } from '@angular/core';
-import {
-  ActivatedRouteSnapshot,
-  CanActivate,
-  Router,
-  RouterStateSnapshot,
-  UrlTree,
-} from '@angular/router';
-import { finalize, Observable, of } from 'rxjs';
+import { CanLoad, Router, UrlTree } from '@angular/router';
+import { Observable, of } from 'rxjs';
 import { Store } from '@ngxs/store';
 
-import { map, mapTo } from 'rxjs/operators';
+import { catchError, map } from 'rxjs/operators';
 import {
   IProfile,
-  LocalStorageState,
   ZWayResponse,
 } from '@store/local-storage/local-storage.state';
 import { ApiService } from '@core/services/api/api.service';
@@ -21,14 +14,14 @@ import { SetUser } from '@store/local-storage/local-storage.actions';
 @Injectable({
   providedIn: 'root',
 })
-export class AuthGuard implements CanActivate {
+export class AuthGuard implements CanLoad {
   constructor(
     private store: Store,
     private router: Router,
     private readonly apiService: ApiService
   ) {}
 
-  canActivate(): Observable<boolean | UrlTree> {
+  canLoad(): Observable<boolean | UrlTree> {
     return this.apiService.send<ZWayResponse<IProfile> | null>('session').pipe(
       map((response) => {
         if (response?.data.id) {
@@ -36,8 +29,10 @@ export class AuthGuard implements CanActivate {
           return true;
         }
         return this.router.createUrlTree(['/firstAccess']);
-      })
+      }),
+      catchError(() => of(this.router.createUrlTree(['/firstAccess'])))
     );
+
     // return this.store.select(LocalStorageState.token).pipe(
     //   mapTo(true)
     //   // map((auth) => {