+++ apps/lite/src/app/core/interceptors/api/api.interceptor.ts
@@ -7,8 +7,8 @@ import {
   HttpRequest,
   HttpResponse,
 } from '@angular/common/http';
-import { Observable } from 'rxjs';
-import { tap } from 'rxjs/operators';
+import { Observable, of } from 'rxjs';
+import { catchError, tap } from 'rxjs/operators';
 import { Store } from '@ngxs/store';
 import { ServerStatus } from '@store/locals/locals.actions';
 import { LocalStorageState } from '@store/local-storage/local-storage.state';
@@ -18,10 +18,7 @@ import { Logout } from '@store/local-storage/local-storage.actions';
 export class ApiInterceptor implements HttpInterceptor {
   constructor(private store: Store) {}
 
-  intercept(
-    request: HttpRequest<unknown>,
-    next: HttpHandler
-  ): Observable<HttpEvent<unknown>> {
+  intercept(request: HttpRequest<unknown>, next: HttpHandler) {
     // let errorsCount = 0;
     // const token = this.store.selectSnapshot(LocalStorageState.token);
     // // this.store.selectSnapshot<string>(AuthState.token);
@@ -35,27 +32,41 @@ export class ApiInterceptor implements HttpInterceptor {
 
     // console.log(request);
     return next.handle(request).pipe(
-      tap({
-        next: (event) => {
-          if (event instanceof HttpResponse) {
-            this.store.dispatch(new ServerStatus(true));
-            // errorsCount = 0;
-          }
-        },
-        error: (err) => {
-          if (err instanceof HttpErrorResponse) {
-            console.log('Error from Inspector', err);
-            if (err.status === 401) {
-              this.store.dispatch(new Logout());
-              console.log('Unauthorized');
-            }
-            if (err.status === 0) {
-              this.store.dispatch(new ServerStatus(false));
-              console.warn('offline');
-            }
-          }
-        },
+      tap((event) => {
+        if (event instanceof HttpResponse) {
+          this.store.dispatch(new ServerStatus(true));
+          // errorsCount = 0;
+        }
       })
+      // catchError((err) => {
+      //   console.warn('http error: ', err.status, err.message);
+      //   if (err.status === 0) {
+      //     this.store.dispatch(new ServerStatus(false));
+      //     console.warn('offline');
+      //   }
+      //   return of(err);
+      // })
+      // tap({
+      //   next: (event) => {
+      //     if (event instanceof HttpResponse) {
+      //       this.store.dispatch(new ServerStatus(true));
+      //       // errorsCount = 0;
+      //     }
+      //   },
+      //   error: (err) => {
+      //     if (err instanceof HttpErrorResponse) {
+      //       console.log('Error from Inspector', err);
+      //       if (err.status === 401) {
+      //         this.store.dispatch(new Logout());
+      //         console.log('Unauthorized');
+      //       }
+      //       if (err.status === 0) {
+      //         this.store.dispatch(new ServerStatus(false));
+      //         console.warn('offline');
+      //       }
+      //     }
+      //   },
+      // })
       // retryWhen((errors) => {
       //     errorsCount++;
       //     console.log({ errorsCount });