+++ CHANGELOG.md
@@ -1,12 +1,20 @@
-## 1.1.4 (2022-02-13)
+## 1.1.5 (2022-02-18)
 
 
 ### Features
 
+* add app version at settings page ([fc4e90a](https://github.com/Xel4ek/Z-Wave-Light/commit/fc4e90a47a90aea454374236220f9a59ad7495f8))
+* add first-access route ([77ee591](https://github.com/Xel4ek/Z-Wave-Light/commit/77ee5918284afe616d6a81bc46dfb2169a0da334))
+* add hideSingleDeviceEvents in settings ([2f7088a](https://github.com/Xel4ek/Z-Wave-Light/commit/2f7088ad96639ff0e0a624b8e08979b45708117e))
+* add login page ([c6af2ae](https://github.com/Xel4ek/Z-Wave-Light/commit/c6af2ae9fb69f38b08e433630b9fba47b512e2b7))
+* add login page ([c27d40b](https://github.com/Xel4ek/Z-Wave-Light/commit/c27d40b03c7d634ad36b8ba6eef4fd0aea1efeb6))
 * add order control in header menu ([d863075](https://github.com/Xel4ek/Z-Wave-Light/commit/d8630757b5c865819489c187d7b8131ddd1f23f7))
+* add personal settings module ([5939244](https://github.com/Xel4ek/Z-Wave-Light/commit/5939244925e8d7cad8253702594c5ccfde290713))
 * CHANGELOG ([b1258f0](https://github.com/Xel4ek/Z-Wave-Light/commit/b1258f0c1aaac92011f0a3d86ee4cc2bf0db7ab1))
 * clear reset filters when route ([7c4c3dd](https://github.com/Xel4ek/Z-Wave-Light/commit/7c4c3dde23bb95399c7d67f6eae16efcaf490aca))
+* protect route login/firstAccess/app ([b7d87a2](https://github.com/Xel4ek/Z-Wave-Light/commit/b7d87a272c444c3108e488047f385d17c232940b))
 * side nav ([b7c1bee](https://github.com/Xel4ek/Z-Wave-Light/commit/b7c1bee35cba419864f352b8647555b113c3cf3d))
+* support multiply profiles ([bdaa4c6](https://github.com/Xel4ek/Z-Wave-Light/commit/bdaa4c61d33ba52a97a5e7a4fb0ff03dc54ef9c6))
 * Tags && filter added in subnav menu ([181f200](https://github.com/Xel4ek/Z-Wave-Light/commit/181f200be0e0714841fdd130622fb078d1392295))
 * versions ([c5b8253](https://github.com/Xel4ek/Z-Wave-Light/commit/c5b8253f3023845af5e39c43eb091cd179d3b494))